Update backtrader to match cryptotrader
There were a few missing calls and parameters. Futhermore, the user can
now specify a period of data to backtest. This can for example be
a datestring looking like (parsed with Date.parse()):

--begin "2014-01-01 GMT+1"

The datetimes shown in the cryptotrader interface are in the user's
local timezone, whereas passing --begin "2014-01-01" will likely result
in the UTC equivalent of that specific date.