Add windows version of talib
Needs the user to change the require-line in talib-sync to
'./talib-win64.node' instead.

node-talib does not work by default on windows platforms. With a bit of
work one can node-gyp the module (which is what this is). And with a
little bit of extra (future) work, the node-talib package can be made
such that it can be installed directly by npm.