Fix mutation and crossover functions
For some magical reason it seemed to work, but probably not always in a
correct way. Class/static functions have now been made static, and only
take in and return a "values" object (non-destructive in the input).

Furthermore, the mutation was kind of big, so the mutations have been
split into two. The bigger mutations are equoal to the mutations that used
to occur before this change, but with a smaller chance.